run: use --import, do not intercept exitCode
If plugins export an `importLoader`, and Module.register exists, then
use that instead of their `loader` export.

process.exitCode became {configurable:false} in
https://github.com/nodejs/node/pull/44711

Can bring back exitCode interception when/if it becomes configurable
again, re https://github.com/nodejs/node/pull/49579

For now, just set it, and then verify it's the expected value, and put
it back to 0 if so.