stack: update for node v20 changes
Some error stack frames string values changed subtly.

Test some things in both CJS and MJS flavors, because this module is
loaded as part of the load process, and then is already cached when
coverage is applied, so we need to run the tests against both to get
full coverage to register.