TypeScript 5.2, use tshy for hybrid builds
This is a massive patch, but only a few things have really changed.

TypeScript upgraded to version 5.2, which has new and more restrictive
behaviors around module loading and ESM/CommonJS dialect emitting.

To work around this, and provide a simpler and more consistent build
experience, tshy is now used to build all modules, the built Test code,
and used in the @tapjs/create-plugin initializer for use in future
plugins.

The dist directories are thus renamed from `mjs` and `cjs` to the more
comprehensible `esm` and `commonjs`. This is a no-op for most users, but
it did mean a lot of edits to import paths in tests.

A shortcoming that was surfaced by the build change showed that it was
possible to add a plugin that can _only_ be loaded by `require()`, and
not by `import()`. This is now detected, and the plugin addition is
prevented in that case.