Do not unnecessarily sort plugin names
The plugin *imports* are already sorted, and we need the pluginNames
array to be in corresponding order, not sorted on its own.

The fix on ccc165ee872d1c7861972a5b0c9ba8ecfcfbea72 was intended to
address https://github.com/tapjs/tapjs/pull/891, but actually did not.
(It did fix a different a subtly related issue, however.)

This addresses the problem by not sorting pluginNames.

Fix: #899