Add --fail-skip/todo opts, better skip reporting
This reports 'no tests found' as the skip reason for any test that does
not have a parent (so, typically, just the root TAP test) and also emits
no tests with a 1..0 plan, even if it lacks a skipReason in the plan.

Also, if `failSkip` is set in the test options (defaulting to
`env.TAP_FAIL_SKIP` in the root test) then any skipped test points will
be treated as failures. Ditto todo tests with `failTodo`.

This makes it possible to run `tap --fail-skip --fail-todo -s todo.txt`
to easily work through a backlog of skipped tests.

Doesn't exactly play nicely with --grep and --only, but it's hard to
know what to do in that case. You're telling tap that you want to fail
if any tests are skipped, and also telling it to skip tests. It seems
reasonable to assume that you don't want it to fail for tests skipped
with --grep or --only, since it should be as if those tests just didn't
*exist*, so probably going to need to figure out another way to signal
this.

Fix: https://github.com/tapjs/tapjs/issues/892