Add --filter-quietly config flag
This makes it possible to use `--grep` and `--only` along with
`--fail-skip`, without making a ton of excessive noise. If you *want*
`--fail-skip` to fail on filtered tests, then you can set
`--no-filter-quietly` to force the skip messages.

Also, it's kind of nicer when filtering tests anyway, because it
provides a way to not have them show a bunch of noisy repetitive skip
messages.