core: move failSkip/failTodo detection to TestBase
The TestPoint class should really be *strictly* a string generator,
having pass/fail change in that function muddies the waters.

Part of TestBase's (admittedly, overly large) job is to determine whether
a test point should be a pass or a failure, so that it can decide whether
to show diagnostics, capture a stack trace, and so on.