pnpm affordance: declare deps, resolve loaders
When using a package manager that does not do a flattened install in
node_modules, the Test class won't be able to load the relevant
dependencies with import() or require(). Also, the loaders may not be
found at the top level (or worse, in the case of ts-node, the *wrong*
version might be found at the top level).

This change resolves all loaders when generating the arguments for test
processes, which may improve startup time just a tad due to node not
having to resolve these things each time, but more importantly, will
ensure that the version which is loaded is the version that the Test
class was built against.

Also, during the default build from the root of the project, run when a
new builtin plugin is added or the Test class is changed, the list of
builtin plugins will be added to src/test/package.json as pinned
dependencies. This required adding an override for nx to know that it's
not actually a relevant dependency cycle, which also means that src/test
can use 'prepare' as its build script target like everything else.

Fix: https://github.com/tapjs/tapjs/issues/915