fix timing issue breaking node --test serialization
The TestMap class was getting confused between the `Test` object and the
`TestBase` object, becuase the `.t` member can't be accessed while the
Test class is still constructing.

The intent was for it to swap over to the t.t once available, but that
wasn't happening properly if multiple a subtests were created in the
same tick as the root TAP object being created.

In any event, it was overly clever. Just slap a pseudorandom key on it
with a known Symbol property, and use that as the key in the diags and
subs maps.