Comb through every combo of plan, promise, end()
When a test function returns a promise, and there's a promise-emitting
assertion like t.resolves() or t.rejects() that triggers a `t.waitOn`
call to occupy the test, it was failing to recognize that it's awaiting
an implicit end, so when the occupying Waiter resolves, it would just
stop and let the process exit, triggering a "test not finished" error in
some cases.

In other cases, if the Waiter was awaited, it _would_ end the test when
the promise returned, but ahead of the resolving Waiter being processed,
resulting in a "test count exceeds plan" error.

This patch adds a test that exercises every combination of plan(), test
fn returning a promise, a pending Waiter, an awaited Waiter, and an
explicit end.

Fix: https://github.com/tapjs/tapjs/issues/937