prevent global/local plugin installation footgun
This makes the tap runner load itself from the current project install
if it's being run from somewhere else, so that plugins etc all work as
expected and stay in sync.

If that's not possible, and the user is attempting to run plugin
commands using a tap executable from some other location, print a
warning, and the command to try if it fails.

Fix: #940