export test-built where node10 tsc can find it
When in `moduleResolution: "node10"` mode, TSC will not resolve package
types using the `exports` field.

In a perfect world, we'd all have long since stopped relying on tools
that are so far out of date with reality, but this world is nothing if
not imperfect.

The solution, thankfully, is relatively simple. Since the built test
class is actually found at `./test-built`, we can just export it as both
`./built` and `./test-built`, and load it from `@tapjs/test/test-built`.

TSC in node10 mode will resolve `./test-built` as a subfolder, find it's
a package, and pull in the `main`, reading the types as CommonJS (and if
that's not what the user is actually building, the types will be subtly
incorrect, but well, they need to be using correct configuration to get
correct types, and a whole heap of stuff will all be in the same
semi-broken state, no way to fix that).

In actual practice, the `./test-built` export will be resolved using the
`"exports"` in package.json, appropriately resolved based on the
import/require conditions.

I'm tempted to _only_ export it as `./test-built`, but figured why have
a breaking change about it. It's only a few lines of json, after all.

Fix: https://github.com/tapjs/tapjs/issues/943

(again)