tapjs/mock: work around ESM hook deadlock
When an ESM hook depends on the main thread to respond, and is
registered prior to another ESM hook being registered, it can deadlock
because the loader freezes the main thread to execute the registration
pseudo-synchronously.

In other words:

- mock loader is registered, now all resolve/load actions will await a
  response from the main thread
- next loader tries to register, using the mock loader (and all previous
  loaders, in fact), *pausing the main thread* to resolve the loader
  module.
- because main thread is paused, @tapjs/mock hook never completes,
  because main thread cannot respond.

The solution here is not really very elegant for solving this problem,
but is a worthwhile optimization in its own right. The client will do
nothing and just fall back to default load/resolve until it gets a
message from the service on the main thread saying that it's been
activated. This message is sent the first time the user calls
`t.mockImport`, since before then there's definitely nothing to mock
anyway.