chore: use --loader in bootstrap
The CI was failing for older node versions, because the bootstrap script
was updated to use `--import` instead of `--loader`, which of course is
not going to work for old node versions.

Could have a switch, but since `--loader` still works, may as well just
leave it using the older style, as long as node < 20 is still supported.