config: resolve paths against config file dir
This also renames `globCwd` to the more appropriate `projectRoot`.

Verify that extension cycle detection doesn't result in false positives
(it was marking the relative path, not the fully resolved path, which
might be different if extended from a different place).

BREAKING CHANGE: any references to a coverage-map, before, or after in
an extension config will have to be updated to be relative to the file
containing the reference, which really is how it should have worked all
along.

Also, since this touched a ton of stuff anyway, formatted everything and
updated to prettier 3.