Fully isolate plugin build actions
This moves the build script into a typical built module, instead of
shipping ts and a js wrapper, so that we don't fail when the project's
TypeScript version is incompatible with tap's.

It also corrects a resolution error, where it was attempting to resolve
the installed plugins from the directory, instead of from a file _in_
the directory. This is a sneaky bug that's hard to test for, because it
works in dev (where all the plugins are reachable from anywhere in the
monorepo) but won't work in an actual project, unless it happens to have
the plugin installed as a dep already. And the entire point of moving
the plugins into the `.tap` folder is so that you don't have to do that.

Fix: https://github.com/tapjs/tapjs/issues/1029