make the implicit-end sigil a keyed symbol
When plugins build outside of the project's node_modules folder, they
can pull in distinct copies of @tapjs/core, resulting in a dual module
hazard.

Surely a better approach could be worked out, but this fixes the issue
for now. The error manifested in a spurious `Error: test end() method
called more than once` being thrown if a root `t.teardown()` is added.

Because both the `process.on('beforeExit')` handler *and* the
`autoend()` behavior triggered by adding a teardown are trying to say
that they're implicit ends, but they're getting the wrong copies of the
symbols to indicate that.

So, instead of a unique symbol, just set it as a `Symbol.for`, so
they'll all find it. Basically it's a global now, but whatever.